Split timeline.js into many files and a new folder
